# spec/api-contract.yaml
openapi: 3.1.0
info:
  title: TTL-Aware BFF API
  version: 1.0.0
  summary: TTL-aware intelligent data-source routing Backend-for-Frontend.
  description: |
    Read-only aggregation tier over two heterogeneous sources:

      * **ODS** — Operational Data Source (gRPC, low latency, current state)
      * **EDS** — Execution Data Source (REST, higher latency, workflow / history / audit)

    Every request is classified, routed through an ordered 11-rule policy chain
    driven by configurable TTLs, source health, required fields and consistency
    requirements, mapped into a single canonical domain model, resolved by an
    explicit source-precedence policy, and returned inside a stable envelope
    carrying freshness and provenance metadata.

    Normative requirements: `spec/requirements.md`.
    Routing semantics: `spec/routing-policy.md`.
    Error semantics: `spec/error-model.md`.

    ## Response envelope

    Every 2xx body is `{ "data": ..., "meta": ... }` (REQ-API-004). `meta` always
    carries the routing decision, the stable routing rule id, the freshness
    verdict, per-field-group provenance, and structured warnings.

    ## Degradation

    `meta.degraded` means the data is old or came from a fallback source — both
    the routing-time rung (`health.primary_unavailable`) and the call-time rung
    (`fallback.primary_failed`) set it, each with a `SOURCE_UNAVAILABLE` warning.
    `meta.partial` means at least one source the request wanted did not answer.
...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
  /resources/{resourceId}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
      - $ref: '#/components/parameters/CorrelationId'
      - $ref: '#/components/parameters/TenantIdHeader'
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/TraceParent'
      - $ref: '#/components/parameters/Consistency'
      - $ref: '#/components/parameters/Fields'
    get:
      tags: [resources]
      operationId: getResource
      summary: Read a resource
      description: |
        Classification `resource_read`. Preferred source OPERATIONAL with
        `ttl: 30s`, `cache_ttl: 5s`, `fallback: execution`, `allow_stale: true`,
        `max_stale: 300s`.

        Typical routing: `ttl.operational.fresh` → OPERATIONAL. When the
        operational record is stale and the EDS can supply the required field
        groups, `ttl.operational.stale` falls back to EXECUTION and the response
        is partial (operational-only groups such as `configuration` and
        `topology` are absent).
      x-requirement:
        - REQ-API-001
        - REQ-API-004
        - REQ-CLS-001
        - REQ-RT-001
        - REQ-TTL-002
        - REQ-EDGE-001
        - REQ-EDGE-002
        - REQ-EDGE-003
      x-request-type: resource_read
      responses:
        '200':
          $ref: '#/components/responses/ResourceOk'
        '206':
          $ref: '#/components/responses/ResourcePartial'
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '405': { $ref: '#/components/responses/MethodNotAllowed' }
        '413': { $ref: '#/components/responses/RequestTooLarge' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/Internal' }
        '502': { $ref: '#/components/responses/UpstreamInvalid' }
        '503': { $ref: '#/components/responses/NoSourceAvailable' }
        '504': { $ref: '#/components/responses/UpstreamTimeout' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Required on every `/api/v1/*` route (REQ-SEC-001).

        Validation (REQ-SEC-002): signature against the JWKS at
        `security.jwt.jwks_url` (or HS256 from `security.jwt.hs256_secret_env`
        in development only), `iss`, `aud`, `exp`, `nbf`, `iat`, with
        `security.jwt.leeway` (default 30s) applied symmetrically. `alg` is
        allow-listed; `none` and algorithm-confusion attempts are rejected. All
        `security.jwt.required_claims` must be present.

        The tenant is resolved from the token claim (`tenant_id` by default) and
        is authoritative. When `X-Tenant-ID` is also supplied it must match the
        claim, otherwise `403 TENANT_MISMATCH` (REQ-SEC-004).

        Authorization is by role → permission mapping per endpoint
        (REQ-SEC-005); see `spec/security.md`.
  parameters:
    ResourceId:
      name: resourceId
      in: path
      required: true
      description: |
        Tenant-scoped resource identifier, the join key between ODS and EDS.
        Validated before anything else touches it: letters, digits and
        `- _ . :` only, at most 128 characters. Anything else is `400`, refused
        before a data source is contacted.
      schema:
        type: string
        pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$'
        minLength: 1
        maxLength: 128
      example: res-1042
...
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      description: |
        Client-supplied correlation id, echoed in the response header and in
        `meta.correlationId`. Generated as a UUIDv4 when absent or malformed
        (REQ-API-010). Propagated to both sources.
      schema:
        type: string
        pattern: '^[A-Za-z0-9._-]{1,128}$'
      example: 2f9c1a7e-3b64-4c2a-9d15-7a0f4b8e1c33
    TenantIdHeader:
      name: X-Tenant-ID
      in: header
      required: false
      description: |
        Optional tenant assertion. When present it MUST equal the JWT tenant
        claim, otherwise `403 TENANT_MISMATCH` (REQ-SEC-004).

        In a `local` or `test` environment running with
        `security.allow_insecure_no_auth: true`, which configuration validation
        permits nowhere else, this header *is* the tenant and no token is needed;
        it defaults to `local` when absent. That is how the compose stack is
        driven.
      schema:
        type: string
        pattern: '^[a-z0-9][a-z0-9-]{0,63}$'
      example: acme
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      description: Caller-side request identifier, logged but not interpreted.
      schema:
        type: string
        maxLength: 128
    TraceParent:
      name: traceparent
      in: header
      required: false
      description: W3C Trace Context parent. Honoured as the parent span and propagated to both sources (REQ-API-011).
      schema:
        type: string
        pattern: '^[0-9a-f]{2}-[0-9a-f]{32}-[0-9a-f]{16}-[0-9a-f]{2}$'
      example: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01
    Consistency:
      name: consistency
      in: query
      required: false
      description: |
        Caller consistency override, honoured **only when it is stricter** than
        the level the endpoint is configured with: a caller may ask for more
        consistency than the endpoint promises, never for less (REQ-CLS-005).

        `strong` forbids stale-serve outright and makes rule
        `consistency.strong_requires_operational` fire, which reads the preferred
        source live with no TTL evaluation at all. It also disables the call-time
        fallback: the caller asked for one specific source's live answer and must
        not silently receive another's.

        It additionally **bypasses the cache read**. `Service.load` calls the
        loader directly rather than going through `GetOrLoad`, so `meta.cache.hit`
        is always `false`; the result is still written back via `Manager.Store`,
        because callers at weaker levels can legitimately use it (REQ-CACHE-006).

        There is no `Cache-Control: no-cache` handling on the request path; this
        query parameter is the only way to raise consistency.
      schema:
        type: string
        enum: [eventual, bounded, strong]
      example: strong
    Fields:
      name: fields
      in: query
      required: false
      description: |
        Comma-separated projection over the canonical field names in
        `FieldGroup`. It is a *routing* input as well as a response filter: the
        field catalogue is what lets the router prove a request needs only one
        source. At most 32 fields; an unknown name is `400 INVALID_REQUEST`.
      schema:
        type: string
      example: status
...
  headers:
    XCorrelationID:
      description: Echo of the request correlation id, or the generated one (REQ-API-010).
      schema: { type: string }
...
  responses:
    ResourceOk:
      description: |
        Resource resolved, complete. `meta.degraded` indicates the data is stale
        or came from a fallback source; a *complete* degraded answer is still a
        `200`. An answer that is missing fields is a `206` instead.
      headers:
        X-Correlation-ID: { $ref: '#/components/headers/XCorrelationID' }
        X-BFF-Source: { $ref: '#/components/headers/XBFFSource' }
        X-BFF-Freshness: { $ref: '#/components/headers/XBFFFreshness' }
        X-BFF-Degraded: { $ref: '#/components/headers/XBFFDegraded' }
        Cache-Control: { $ref: '#/components/headers/CacheControlOut' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ResourceEnvelope' }
          examples:
            fresh: { $ref: '#/components/examples/ResourceFresh' }
            staleDegraded: { $ref: '#/components/examples/ResourceStaleDegraded' }
    InvalidRequest:
      description: Malformed path parameter, out-of-range `limit`, malformed cursor, or missing tenant.
      headers:
        X-Correlation-ID: { $ref: '#/components/headers/XCorrelationID' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            badLimit: { $ref: '#/components/examples/ErrInvalidRequest' }
            tenantMissing: { $ref: '#/components/examples/ErrTenantMissing' }
    Unauthenticated:
      description: Missing, malformed or invalid bearer token (REQ-SEC-002).
      headers:
        X-Correlation-ID: { $ref: '#/components/headers/XCorrelationID' }
        WWW-Authenticate:
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            expired: { $ref: '#/components/examples/ErrUnauthenticated' }
    Forbidden:
      description: Authenticated but not permitted, or tenant assertion mismatch (REQ-SEC-004, REQ-SEC-005).
      headers:
        X-Correlation-ID: { $ref: '#/components/headers/XCorrelationID' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            forbidden: { $ref: '#/components/examples/ErrForbidden' }
            tenantMismatch: { $ref: '#/components/examples/ErrTenantMismatch' }
    ResourcePartial:
      description: |
        The body does not carry everything the request asked for. For
        `resource_read` this is the **fallback** case: the operational source was
        unavailable — pre-flight (`health.primary_unavailable`) or at call time
        (`fallback.primary_failed`) — and the execution source answered instead.
        The EDS holds no `configuration`, `owner`, `metrics` or `topology`, so
        those fields are absent and `meta.partial` is `true`.

        Absent fields are omitted, never null-filled, and a `PARTIAL_DATA`
        warning says so explicitly.
      headers:
        X-Correlation-ID: { $ref: '#/components/headers/XCorrelationID' }
        X-BFF-Source: { $ref: '#/components/headers/XBFFSource' }
        X-BFF-Freshness: { $ref: '#/components/headers/XBFFFreshness' }
        X-BFF-Degraded: { $ref: '#/components/headers/XBFFDegraded' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ResourceEnvelope' }
          examples:
            preflightFallback: { $ref: '#/components/examples/ResourcePreflightFallback' }
            ttlMissToExecution: { $ref: '#/components/examples/ResourceFallbackExecution' }
    NotFound:
      description: |
        Every consulted source reports the resource absent (REQ-EDGE-019a).
        Negatively cached for `cache.negative_ttl`. An existing resource with an
        empty collection is `200`, not `404`.
      headers:
        X-Correlation-ID: { $ref: '#/components/headers/XCorrelationID' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            notFound: { $ref: '#/components/examples/ErrNotFound' }
    MethodNotAllowed:
      description: The path exists but the method is not GET (REQ-API-001).
      headers:
        Allow: { $ref: '#/components/headers/Allow' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    RequestTooLarge:
      description: Request exceeded `server.max_body_bytes` or header limits (REQ-SEC-010).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...